Fix WordEdit
