+++ lib/components/quiz/unsolved_content.dart
@@ -103,22 +103,17 @@ class QuizUnsolvedContentState extends ConsumerState<QuizUnsolvedContent> {
             print('solvedQuizIds');
             print(solvedQuizIds);
             print(solvedQuizIds.length);
+            // loadedQuizIds から solvedQuizIdsを取り除くことで、
+            // 読み込まれた問題（ loadedQuizIds ）のうち、実際に画面に表示されている問題のIDを算出する （読み込まれた問題 - 解いた問題のID （solvedQuizIdsProvider））
             loadedQuizIds
                 .removeWhere((int quizId) => solvedQuizIds.contains(quizId));
             print('loadedQuizIds');
             print(loadedQuizIds);
             print(loadedQuizIds.length);
-
-            // 読み込まれた問題（loadedQuizIds）のうち、実際に画面に表示されている問題のIDを算出する （読み込まれた問題 - 解いた問題のID （solvedQuizIdsProvider））
-            //loadedQuizIds
-            //    .removeWhere((int quizId) => solvedQuizIds.contains(quizId));
-            //print('displayedQuizIDs');
-            //print(loadedQuizIds);
-            // 画面に表示されている問題がなければ、次の問題を読み込むために親（reviewやdrillのscreen）に通知する。
-            //
-            // 10の倍数の解答数でリロードする
+            // 画面に表示されている問題がないか、10の倍数の解答数で、次の問題10件を読み込み始める。
             if (solvedQuizIds.length % 10 == 0 || loadedQuizIds.isEmpty) {
               print("LoadingNext");
+              // 次の問題を読み込むために親（reviewやdrillのscreen）に通知する。
               LoadingUnsolvedQuizzesNotification(true).dispatch(context);
             }
             setState(() {
